backend/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dineroclaro-backend"
version = "0.1.0"
description = "DineroClaro FastAPI service"
authors = [{ name = "DineroClaro" }]
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.29.0",
  "pydantic-settings>=2.2.1",
  "sqlalchemy>=2.0.29",
  "psycopg2-binary>=2.9.9",
  "supabase>=2.3.4",
  "openai>=1.14.3",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100
